user state/team state
